bin: split main logic into `index.js` file
